This message was manufactured by g4svn to create branch 09-04-ref-05_branch
git-svn-id: svn+ssh://svn.cern.ch/reps/geant4/branches/geant4/_symbols/09-04-ref-05_branch@50133 82ad68bf-922b-4865-b672-c5fc3b8c97f6

+++ examples/extended/eventgenerator/HepMC/MCTruth/src/MCTruthManager.cc
@@ -275,8 +275,8 @@ void MCTruthManager::printTree(HepMC::GenParticle* particle, G4String offset)
   {
     G4String deltaoffset = "";
 
-    G4int curr = std::fmod((*it)->barcode(),10000000);
-    G4int part = std::fmod(particle->barcode(),10000000);
+    G4int curr = std::fmod(double((*it)->barcode()),10000000.);
+    G4int part = std::fmod(double(particle->barcode()),10000000.);
     if( curr != part )
       {
         deltaoffset = " | ";
